+++ NEWS.md
@@ -4,5 +4,8 @@
 ## Minor changes
 * Fixed some types in the documentation
 * New arguments to simulate.plcp_multi; save_folder, etc
+* 'simulate.pclp' new argument 'batch_progress' enables showing progress when doing
+multiple simulations.
+